NF: PersistentCache with @memoize and @memoize_path
Many functions in dandi-cli rely on analysis of metadata from .nwb files.
Getting it takes time, and we might want to do that in multiple independent
runs of dandi.  PersistentCache, based on joblib.Memory, provides convenience
such as @memoize_path to just decorate functions where the first argument is
the path, and then memoize the returned value (oh, do we need to support generators as well??)