ENH+BF: upload/download - use uploaded_mtime, collect errors during upload
but do not use those lists of errors ATM, since there is only one place
in the code to get them and it is the last step and they are then included in pyout
output anyways.

Also fixed naming of devel_debug which hindered my demo.

upload_mtime is now used instead of girder File ctime, since regular
user cannot set it.