ENH: upload - add locking of the dandiset
It locks the entire dandiset for the entire operation (loop across all
files), even if no actual upload would happen (e.g. if all files
already uploaded and recent).

It has a few TODO items added in the code which ideally should be
resolved before we proceed with merging this PR.