RF: do not upload dandiset.yaml
With RFings in master we can screw it up, and also web ui is fragile
to wrong metadata records atm and there is no validation yet.  Since
we already have a header saying that all changes would be lost, I think the
best course of action is just to not upload it at all!