Use timed wait in publish test
I think the flaky tests we have been seeing are being caused by multiple
sequential celery tasks. Creating a new dandiset and uploading an asset
both trigger the validation task, but the first one is guaranteed to
fail. I think the test was noticing that first failure and quitting
immediately, instead of waiting the extra time required for the second
task to validate correctly.