RF: attempt nwb metadata extraction only on .nwb, if fails -- warning
Maintaining prior behavior of re-raising the exception if allow_any_path is
False.  Otherwise -- and regardless if allow_any_path is True or not (thus
would not re-raise for non .nwb files even if not allow_any_path) - provide default
metadata record.

For upload it should be fine AFAIK even with the change  of behavior since we first
use allow_any_path to select .nwb files.

Closes #730