Specify DJANGO_DANDI_VALIDATION_JOB_INTERVAL
https://github.com/dandi/dandi-api/pull/538 is adding a scheduled job
which checks for Versions that need validation and validates them. The
default wait interval is 60 seconds, but it is configurable through an
environment variable, `DJANGO_DANDI_VALIDATION_JOB_INTERVAL`. Set this
environment variable to `5` so that the validation job runs every 5
seconds, which should be quick enough for `wait_until_valid`.