BF: (optionally) guard download_generator to not propagate errors
This way instead of dumping all tracebacks at the end, some times for benign
FileExistsError -- they will be displayed neatly in pyout and also output in
detail in the log by default in the case of pyout output handling. But in
"debug" mode we will not do anything like that.

Closes https://github.com/dandi/dandi-cli/issues/1007