organize - sanitize ; as well within the entity value
User reported filenames with ; in them.  Although I do not think
it caused any trouble, seeing such character in the filename is
quite "surprising" to say the least, and since ; has meaning in
POSIX as separator between different commands I think best to be avoided.
Also it is not part of alphanumerics of BIDS.

But with that I wonder if we should add other punctuation to be
subject for sanitization -- e.g.  ,[](){}?