BF: fail upload only for validation ERRORs and not necessarily WARNINGs or HINTs + log all errors
Reported by a user in slack to be unable to upload a file, although validation did not
report any errors.  Apparently we did not log errors and it was impossible to figure out
what is going on.  So I have also added logging of those validation errors, without any custom
serialization into a string so we log the most complete (and possibly the most useful since
provides information on paths within file etc) representation.