Move doc/design under docs/ where the rest of docs reside
in dandi-archive we have them under doc/ so I guess that is why had them under doc/
here as well, but IMHO there is no reason to split documentation here into two folders.
If we are to unify, we might want to rename doc/ to docs/ in dandi-archive as well.

Alternative -- move everything under docs/* into doc/ but we might need to tune up
more (RTD etc). WDYT?