RF+BF: Explicitly open h5py.File as read-only where we read it
ref: https://github.com/dandi/helpdesk/discussions/95#discussioncomment-5687405

apparently we started to fail to extract metadata due to some warning being
spitted out that file is not open for writing whenever we do not even expect it
to open for writing.  Not yet entirely sure what was happening there since user
reported that the issue was resolved by upgrade of h5py. But here it is
worth addressing the warning anyways:

    H5pyDeprecationWarning: The default file mode will change to 'r' (read-only) in h5py 3.0. To suppress this warning, pass the mode you need to h5py.File(), or set the global default h5.get_config().default_file_mode, or set the environment variable H5PY_DEFAULT_READONLY=1. Available modes are: 'r', 'r+', 'w', 'w-'/'x', 'a'. See the docs for details. with open_readable(filepath) as f, h5py.File(f) as fp: