ENH: provide descriptive assertion error on attempts_allowed_or_not
Somehow, with some prior version of the code, but we did
get this assertion triggered as shown in
https://github.com/dandi/dandi-cli/issues/1549#issuecomment-2537027028
although it remains unclear how that could have happened if type checking
is correct since we must not end up there with None or even 0.

Unfortunately with a regular assert we just get an announcement of the fact
and no details.  After this change we hope to be able to see more details
which might hint on the original issue.