Allow older clients to upload as long as it is a version known to be upgradable
ATM server DOES NOT CARE! So we already allow for assets being created by other clients
with outdated schema etc. They fail validation unless could be upgraded!

Changes might need to be adjusted based on the name for the /info record
field as would be decided in https://github.com/dandi/dandi-archive/issues/2624
or PR to solve it